requirements.txt
django-ninja==1.1.0
django-environ==0.11.2
itsdangerous==2.1.2